backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
qdrant-client==1.7.0
openai==1.12.0
tiktoken==0.5.2
gitpython==3.1.41
pygments==2.17.2
pytest==7.4.3
python-dotenv==1.0.1